Create components to use in the Gifs section.
Create 'home-page' component in the 'pages' folder to contain the structure of the Gifs section.
Create 'search-box' component to search Gifs.
Create 'card-list' component to show the searched Gifs.
Add the created components to the 'declarations' section of the 'gifs.module.ts' file.
Add the 'home-page' componente to the 'exports' section of the 'gifs.module.ts' file so it can be used outside the module.
Add the 'home-page' component in the 'app.component.html' file.